requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.1.3
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.1.2
gradio_client==0.7.0
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.19.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
ml-dtypes==0.2.0
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tomlkit==0.12.0
toolz==0.12.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
